requirements.txt
Django==4.0.6
django-crispy-forms==1.14.0
PyMySQL==1.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
drf-yasg==1.21.5
Pillow==9.5.0